backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx>=0.2.4",
    "fastapi>=0.118.0",
    "firebase-admin>=7.1.0",
    "google>=3.0.0",
    "langchain-google-genai>=2.1.12",
    "langchain-pinecone>=0.2.12",
    "langgraph>=0.6.8",
    "pinecone>=7.3.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.4",
    "pymupdf4llm>=0.0.27",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "pytest-mock",
    "respx",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]